refactor: 도메인 검증 부분 다시 복구
- repository 를 사용하지 않는 검증은 굳이 Validator로 뺄 필요가 없음
